feat: wip WASM support for the WebRTC transport
