+++ p2p/transport/webrtc/transport_test.go
@@ -1,3 +1,6 @@
+//go:build !js
+// +build !js
+
 package libp2pwebrtc
 
 import (
